Rename main.js to markdownProofing.js src and test files
