+++ src/lib/markdownProofing.js
